Clean up and reorganise CoE types and functionality (#353)
* More sensible names

* Rename a bunch of stuff, flatten a few structs

* Refactor mailbox module with CoE module inside

* Refactor even more CoE stuff

* Move SubDevice CoE stuff into its own file
